moved files for doc generation into script folder adjusted scripts so root version isn't longer necessary
