--- index.html
@@ -1,5 +1,0 @@
-<html>
-<meta http-equiv="refresh" content="0; URL='README.html'" />
-<body>
-</body>
-</html>
