+++ scripts/docs/docfx.json
@@ -6,10 +6,10 @@
         {
           "files": [ "**/*.sln", "**/*.cs" ],
           "exclude": [ "**/bin/**", "**/obj/**" ],
-          "src": "Assets"
+          "src": "../../Assets"
         }
       ],
-      "dest": "obj/api"
+      "dest": "../../obj/api"
     }
   ],
   "build": {
@@ -17,15 +17,17 @@
       {
         // generate html for api docs
         "files": [ "**/*.yml"],
-        "src": "obj/api", // temp meta data folder from previous step
-        "dest": "api"
+        "src": "../../obj/api", // temp meta data folder from previous step
+        "dest": "./api"
       },
       {
         // generate html for conceptual docs
         "files": ["*.md", "*.yml", //root dir md and toc files
                 "Documentation/**/*.md", "Assets/**/*.md",
 				"Documentation/**/*.yml", "Assets/**/*.yml" // conceptual docs subfolders toc yml files
-                ]
+                ],
+		"src": "../..",
+		"dest": "./"
       }
     ],
     "resource": [
@@ -33,10 +35,13 @@
       {
         // website related resource files
         "files": ["index.html", "web/**"],
+		"dest": "./"
       },
       {
         // resources documentation is referring to (eg images, videos, audio, etc)
-        "files": [ "Documentation/Images/**" ],
+		"files": ["**/*.*"],
+		"src": "../../Documentation/Images",
+		"dest": "./Documentation/Images"
       },
     ],
     "template": [
@@ -49,8 +54,8 @@
       "_appTitle": "Mixed Reality Toolkit Documentation",
       "_enableSearch": true,
       "_disableNavbar": false,
-      "_appLogoPath": "Documentation/Images/mrt_logo_icon.png",
-      "_appFaviconPath": "Documentation/Images/favicon.ico",
+      "_appLogoPath": "../../Documentation/Images/mrt_logo_icon.png",
+      "_appFaviconPath": "../../Documentation/Images/favicon.ico",
       // section contribute start
 	  // remove this section if you want to restore default behavior which publishes a version of the docs that keeps the "improve this doc" link pointing to the publishing branch
       "_gitContribute": {
@@ -60,7 +65,7 @@
 	  // section contribute end
     },
     "markdownEngineName": "markdig",
-    "dest": "doc",
+    "dest": "./doc",
     "xrefService": [ "https://xref.docs.microsoft.com/query?uid={uid}" ] //msdn xref service for resolving crefs
   }
 }
