+++ scripts/docs/index.html
@@ -1,0 +1,5 @@
+<html>
+<meta http-equiv="refresh" content="0; URL='README.html'" />
+<body>
+</body>
+</html>
