+++ scripts/docs/web_root/index.html
@@ -1,0 +1,5 @@
+<html>
+<meta http-equiv="refresh" content="0; URL='version/releases/2.4.0/README.html'" />
+<body>
+</body>
+</html>
